demo/rosetta_data_integration/requirements.txt
google-cloud-spanner